requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.8.0
scipy>=1.9.0
matplotlib>=3.5.0
Pillow>=9.0.0

# Optional: For advanced depth estimation
# timm>=0.9.0  # For DPT models

# Utility
tqdm>=4.64.0
pathlib
argparse

# Development (optional)
pytest>=7.0.0
jupyter>=1.0.0
